(PC-18090)[PRO] feat: display image in collective offers list
